- made shiny an optional dependency - import exp_shiny directly into ExposedDF
