+++ pyproject.toml
@@ -18,7 +18,7 @@ plotnine = ">=0.12.4"
 joblib = ">=1.3.2"
 scipy = ">=1.11.4"
 matplotlib = ">=3.8.2"
-shiny = ">=0.6.0"
+shiny = {version = ">=0.7.0", optional = true}
 great-tables = ">=0.2.0"
 
 [tool.poetry.dev-dependencies]
